contributors/i18n/submitted/css3-syntax/i18n-css-syntax-bytes-004.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>CSS3 Syntax, input byte stream: @charset declaration (lowercase value)</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-syntax/#input-byte-stream'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='http dom'>
<meta name="assert" content="An @charset rule with a lowercase value sets the encoding of a css stylesheet.">
<style type='text/css'>
.test div { width:50px; }

p { color: gray; margin-top: 3em; font-size: 80%; }
</style>
<link rel="stylesheet" type="text/css" href="support/charset15lcvalue.css">
</head>
<body>



<div class="test"><div id="box" class="ÃœÃ€Ãš">&#xA0;</div></div>


<!--Notes:
In the style sheet, the CSS selector identifies a class name with the following byte sequence: c3 bd c3 a4 c3 a8.  Read as UTF-8 this gives the characters ýäè.  Read as ISO 8859-1, this gives Ã½Ã¤Ã¨. Read as ISO 8850-15, this gives ÃœÃ€Ãš.

This HTML page is served as UTF-8 and uses a different byte sequence for the class name (c3 83 c5 93 c3 83 e2 82 ac c3 83 c5 a1) that equates to the character sequence ÃœÃ€Ãš.  

The class ids will therefore only match if the CSS is read as ISO 8859-15.

The CSS file encoding is established as ISO 8859-15 using only an @charset rule, with the name of the encoding in lowercase.

-->

<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 100);
}, " ");
</script>

<div id='log'></div>

<p class="ÃœÃ€Ãš"><span><span id="debug">Text as read from the style sheet: </span><br />(If blank, the style sheet hasn't been read at all.)</span></p>
</body>
</html>
